refactor(agent): App 渲染函数下沉到各自 App，消除 runtime 反向依赖
把 HN / IT之家 / QQ 专属的屏幕渲染从共享的 runtime/context/context-message-factory.ts
下沉到各自 App 目录（hn-screen.ts / ithome-screen.ts / qq-message-render.ts），
让 runtime 核心层不再反向 import apps/* 与 napcat。factory 只留通用消息构造器。
纯代码搬移，行为与序列化输出完全一致；对应单测随之拆分，补齐 QQ 私聊显示名回退单测；
顺带删除 6 个无调用方的死函数。
