+++ apps/agent/src/agent/apps/hn/tools/glance-hn.tool.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 import { ZodToolComponent, type ToolExecutionResult, type ToolKind } from "@kagami/agent-runtime";
-import { renderHnFrontPageContent } from "../../../runtime/context/context-message-factory.js";
+import { renderHnFrontPageContent } from "../hn-screen.js";
 import type { RootAgentEffect } from "../../../runtime/effect/root-agent-effect.js";
 import type { HnReader } from "../hn-reader.js";
