+++ apps/agent/src/agent/apps/hn/tools/open-hn-user.tool.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 import { ZodToolComponent, type ToolExecutionResult, type ToolKind } from "@kagami/agent-runtime";
-import { renderHnUserContent } from "../../../runtime/context/context-message-factory.js";
+import { renderHnUserContent } from "../hn-screen.js";
 import type { RootAgentEffect } from "../../../runtime/effect/root-agent-effect.js";
 import type { HnReader } from "../hn-reader.js";
