+++ apps/agent/src/agent/apps/hn/tools/search-hn.tool.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 import { ZodToolComponent, type ToolExecutionResult, type ToolKind } from "@kagami/agent-runtime";
-import { renderHnSearchContent } from "../../../runtime/context/context-message-factory.js";
+import { renderHnSearchContent } from "../hn-screen.js";
 import type { RootAgentEffect } from "../../../runtime/effect/root-agent-effect.js";
 import type { HnReader } from "../hn-reader.js";
