+++ apps/agent/src/agent/runtime/context/context-message-factory.ts
@@ -1,22 +1,7 @@
 import type { AsyncTaskCompletion } from "@kagami/agent-runtime";
 import type { Event } from "../event/event.js";
 import type { LlmContentPart, LlmMessage } from "../../../llm/types.js";
-import {
-  renderSupportedMessageSegments,
-  type NapcatReceiveMessageSegment,
-} from "../../../napcat/application/napcat-gateway/shared.js";
-import type {
-  NapcatGroupMessageData,
-  NapcatPrivateMessageData,
-} from "../../../napcat/application/napcat-gateway.service.js";
 import { renderServerStaticTemplate } from "@kagami/server-core/common/runtime/read-static-text";
-import type { HnFeed } from "../../apps/hn/client/firebase.js";
-import type {
-  HnGlanceResult,
-  HnThreadResult,
-  HnSearchResult,
-  HnUserResult,
-} from "../../apps/hn/hn-reader.js";
 
 const BEIJING_TIME_ZONE = "Asia/Shanghai";
 
@@ -161,329 +146,10 @@ export function createWebSearchInstructionMessage(question: string): UserMessage
   );
 }
 
-type IthomeArticleListInput = {
-  displayName: string;
-  mode: "latest" | "new";
-  hiddenNewCount: number;
-  articles: Array<{
-    id: number;
-    title: string;
-    url: string;
-    publishedAt: Date;
-    rssSummary: string;
-  }>;
-};
-
-export function renderIthomeArticleListContent(input: IthomeArticleListInput): string {
-  return renderServerStaticTemplate(import.meta.url, "context/ithome-article-list.hbs", {
-    displayName: input.displayName,
-    isNewMode: input.mode === "new",
-    hiddenNewCount: input.hiddenNewCount,
-    articles: input.articles.map(article => ({
-      ...article,
-      publishedAtText: formatDateTime(article.publishedAt),
-    })),
-  });
-}
-
-export function createIthomeArticleListMessage(input: IthomeArticleListInput): UserMessage {
-  return createUserMessage(renderIthomeArticleListContent(input));
-}
-
-type IthomeArticleDetailInput = {
-  title: string;
-  url: string;
-  publishedAt: Date;
-  content: string;
-  contentSource: "article_content" | "rss_summary";
-  truncated: boolean;
-  maxChars: number;
-};
-
-export function renderIthomeArticleDetailContent(input: IthomeArticleDetailInput): string {
-  return renderServerStaticTemplate(import.meta.url, "context/ithome-article-detail.hbs", {
-    title: input.title,
-    url: input.url,
-    publishedAtText: formatDateTime(input.publishedAt),
-    content: input.content.trim(),
-    fallbackToSummary: input.contentSource === "rss_summary",
-    truncated: input.truncated,
-    maxChars: input.maxChars,
-  });
-}
-
-export function createIthomeArticleDetailMessage(input: IthomeArticleDetailInput): UserMessage {
-  return createUserMessage(renderIthomeArticleDetailContent(input));
-}
-
-// === Hacker News App 屏幕渲染 ===
-// 领域模型已在 HnReader 里清洗过（htmlToPlainText 去标签 + 软化尖括号），
-// 所以这里直接拼进 <hn_*> XML 段落是安全的——HN 文本无法伪造闭合标签越狱。
-
-const HN_FEED_LABEL: Record<HnFeed, string> = {
-  top: "热榜",
-  new: "最新",
-  best: "最佳",
-  ask: "Ask HN",
-  show: "Show HN",
-  job: "招聘",
-};
-
-export function renderHnFrontPageContent(result: HnGlanceResult): string {
-  const lines = [`<hn_front_page feed="${HN_FEED_LABEL[result.feed]}">`];
-  if (result.stories.length === 0) {
-    lines.push("（这个榜单暂时没拉到内容）");
-  }
-  result.stories.forEach((story, index) => {
-    const meta = [
-      story.score !== null ? `${story.score} 分` : null,
-      story.descendants !== null ? `${story.descendants} 评论` : null,
-      story.by ? `by ${story.by}` : null,
-      story.domain,
-      story.postedAt ? formatDateTime(story.postedAt) : null,
-    ]
-      .filter(Boolean)
-      .join(" · ");
-    lines.push(`${index + 1}. [id=${story.id}] ${story.title}`);
-    if (meta) {
-      lines.push(`   ${meta}`);
-    }
-  });
-  lines.push("</hn_front_page>");
-  return lines.join("\n");
-}
-
-export function renderHnThreadContent(result: HnThreadResult): string {
-  const lines = [`<hn_thread id="${result.id}">`];
-  lines.push(result.title ?? "(无标题)");
-  const head = [
-    result.by ? `by ${result.by}` : null,
-    result.domain,
-    result.postedAt ? formatDateTime(result.postedAt) : null,
-  ]
-    .filter(Boolean)
-    .join(" · ");
-  if (head) {
-    lines.push(head);
-  }
-  if (result.url) {
-    lines.push(`链接：${result.url}`);
-  }
-  if (result.selfText) {
-    lines.push("", result.selfText);
-  }
-  lines.push("", `--- 讨论（${result.shownRootComments}/${result.totalRootComments} 条主楼）---`);
-  if (result.comments.length === 0) {
-    lines.push("（还没有评论）");
-  }
-  for (const comment of result.comments) {
-    const indent = comment.depth > 1 ? "    " : "";
-    const replyHint = comment.replyCount > 0 ? `（${comment.replyCount} 回复）` : "";
-    const author = comment.author ?? "(匿名)";
-    lines.push(`${indent}- ${author}${replyHint}：${comment.text || "（空）"}`);
-  }
-  if (result.truncated) {
-    lines.push("", "（讨论已截断，还有更多评论没展开）");
-  }
-  lines.push("</hn_thread>");
-  return lines.join("\n");
-}
-
-export function renderHnSearchContent(result: HnSearchResult): string {
-  const sortLabel = result.sort === "date" ? "按时间" : "按热度";
-  const lines = [`<hn_search query="${result.query}" sort="${sortLabel}">`];
-  if (result.hits.length === 0) {
-    lines.push("（没搜到相关内容）");
-  }
-  result.hits.forEach((hit, index) => {
-    const meta = [
-      hit.kind === "comment" ? "评论" : "帖子",
-      hit.points !== null ? `${hit.points} 分` : null,
-      hit.numComments !== null ? `${hit.numComments} 评论` : null,
-      hit.author ? `by ${hit.author}` : null,
-      hit.domain,
-      hit.postedAt ? formatDateTime(hit.postedAt) : null,
-    ]
-      .filter(Boolean)
-      .join(" · ");
-    const head = hit.title ?? hit.snippet ?? "(无标题)";
-    lines.push(`${index + 1}. [id=${hit.id}] ${head}`);
-    lines.push(`   ${meta}`);
-    if (hit.title && hit.snippet) {
-      lines.push(`   ${hit.snippet}`);
-    }
-  });
-  lines.push("</hn_search>");
-  return lines.join("\n");
-}
-
-export function renderHnUserContent(result: HnUserResult): string {
-  const lines = [`<hn_user name="${result.username}">`];
-  if (!result.found) {
-    lines.push(`没找到用户 ${result.username}。`);
-    lines.push("</hn_user>");
-    return lines.join("\n");
-  }
-  const meta = [
-    result.karma !== null ? `karma ${result.karma}` : null,
-    result.createdAt ? `注册于 ${formatDateTime(result.createdAt)}` : null,
-  ]
-    .filter(Boolean)
-    .join(" · ");
-  if (meta) {
-    lines.push(meta);
-  }
-  if (result.about) {
-    lines.push("", result.about);
-  }
-  lines.push("", "--- 近期发言 ---");
-  if (result.recent.length === 0) {
-    lines.push("（最近没有发言）");
-  }
-  for (const item of result.recent) {
-    const kind = item.kind === "comment" ? "评论" : "帖子";
-    const when = item.postedAt ? formatDateTime(item.postedAt) : "";
-    const body = item.title ?? item.snippet ?? "(无内容)";
-    lines.push(`- [${kind}${when ? ` ${when}` : ""}] ${body}`);
-  }
-  lines.push("</hn_user>");
-  return lines.join("\n");
-}
-
 export function createMessagesFromEvent(_event: Event): UserMessage[] {
   // 目前没有任何事件类型需要渲染成上下文消息：notification / story_recall 由 session
   // 直接装配成 <notification> / <story_recall> 消息追加（不是 event 类 ContextItem），
   // wake 是纯唤醒。保留此入口是为 event→ContextItem 渲染路径留口子——未来若有需要直接
   // 进上下文的新事件类型，在这里加分支即可。
   return [];
-}
-
-export function renderMergedGroupMessagesContent(
-  messages: NapcatGroupMessageData[],
-): string | null {
-  if (messages.length === 0) {
-    return null;
-  }
-  return messages.map(message => renderGroupMessagePlainText(message)).join("\n\n");
-}
-
-export function createMergedGroupMessagesMessage(
-  messages: NapcatGroupMessageData[],
-): UserMessage | null {
-  const content = renderMergedGroupMessagesContent(messages);
-  return content === null ? null : createUserMessage(content);
-}
-
-export function renderMergedPrivateMessagesContent(
-  messages: NapcatPrivateMessageData[],
-): string | null {
-  if (messages.length === 0) {
-    return null;
-  }
-  return messages.map(message => renderPrivateMessagePlainText(message)).join("\n\n");
-}
-
-export function createMergedPrivateMessagesMessage(
-  messages: NapcatPrivateMessageData[],
-): UserMessage | null {
-  if (messages.length === 0) {
-    return null;
-  }
-
-  return createUserMessage(
-    messages.map(message => renderPrivateMessagePlainText(message)).join("\n\n"),
-  );
-}
-
-export function renderGroupMessagePlainText(input: {
-  nickname: string;
-  userId: string;
-  rawMessage: string;
-  messageSegments?: NapcatReceiveMessageSegment[];
-  messageId?: number | null;
-}): string {
-  return renderQqMessagePlainText({
-    displayName: input.nickname,
-    userId: input.userId,
-    rawMessage: input.rawMessage,
-    messageSegments: input.messageSegments,
-    messageId: input.messageId,
-  });
-}
-
-export function renderPrivateMessagePlainText(input: {
-  nickname: string;
-  remark: string | null;
-  userId: string;
-  rawMessage: string;
-  messageSegments?: NapcatReceiveMessageSegment[];
-  messageId?: number | null;
-}): string {
-  return renderQqMessagePlainText({
-    displayName: formatPrivateChatDisplayName(input),
-    userId: input.userId,
-    rawMessage: input.rawMessage,
-    messageSegments: input.messageSegments,
-    messageId: input.messageId,
-  });
-}
-
-export function formatPrivateChatDisplayName(input: {
-  nickname: string;
-  remark: string | null;
-  userId: string;
-}): string {
-  const remark = input.remark?.trim();
-  if (remark) {
-    return remark;
-  }
-
-  const nickname = input.nickname.trim();
-  if (nickname) {
-    return nickname;
-  }
-
-  return input.userId;
-}
-
-function renderQqMessagePlainText(input: {
-  displayName: string;
-  userId: string;
-  rawMessage: string;
-  messageSegments?: NapcatReceiveMessageSegment[];
-  messageId?: number | null;
-}): string {
-  const renderedMessage = renderQqMessageBody(input);
-  return renderServerStaticTemplate(import.meta.url, "context/qq-message.hbs", {
-    nickname: input.displayName,
-    userId: input.userId,
-    messageBody: renderedMessage,
-    // 暴露 QQ message_id 作为「回复哪条」的句柄；缺失时模板不渲染 id 属性。
-    messageId: input.messageId ?? null,
-  });
-}
-
-function renderQqMessageBody(input: {
-  rawMessage: string;
-  messageSegments?: NapcatReceiveMessageSegment[];
-}): string {
-  const segments = input.messageSegments ?? [];
-  if (segments.length === 0) {
-    return input.rawMessage.trim();
-  }
-
-  const rendered = renderSupportedMessageSegments(segments);
-  return rendered;
-}
-
-function formatDateTime(value: Date): string {
-  return new Intl.DateTimeFormat("zh-CN", {
-    timeZone: BEIJING_TIME_ZONE,
-    year: "numeric",
-    month: "numeric",
-    day: "numeric",
-    hour: "2-digit",
-    minute: "2-digit",
-    hour12: false,
-  }).format(value);
 }
